backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
websockets==13.1
requests==2.32.3
faster-whisper
soundfile
numpy
pyttsx3
edge-tts